renamed land data generator script
